# nodeGIS/api.yaml
openapi: 3.1.0
info:
  title: 默认模块
  description: ''
  version: 1.0.0
...
paths:
...
  /friends:
    get:
      summary: 获取好友列表
      deprecated: false
      description: ''
      tags:
        - 好友 (Friends)
      parameters: []
      responses:
        '200':
          x-apifox-name: 成功的好友列表
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Friend'
                uniqueItems: true
                minItems: 10
                maxItems: 20
          headers: {}
          x-apifox-ordering: 0
        '401':
          x-apifox-name: 未认证（Token 无效或缺失）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: UNAUTHENTICATED
                message: 需要提供有效的认证Token
          headers: {}
          x-apifox-ordering: 1
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 2
      security:
        - jwt: []
          x-apifox:
            required: true
            schemeGroups:
              - id: 2E3vUtLwdMRJwtG6NLo0V
                schemeIds:
                  - jwt
            use:
              id: 2E3vUtLwdMRJwtG6NLo0V
              configs:
                jwt:
                  authConfigs:
                    x-apifox:
                      token: >-
                        eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw
      x-apifox-folder: 好友
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-380410910-run
    post:
      summary: 添加好友
      deprecated: false
      description: ''
      tags:
        - 好友 (Friends)
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequest'
      responses:
        '201':
          x-apifox-name: 好友添加成功
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Friend'
          headers: {}
          x-apifox-ordering: 0
        '400':
          x-apifox-name: 无效请求（例如：请求体验证失败）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VALIDATION_ERROR
                message: 用户 ID 不能为空
                details:
                  - field: userId
                    issue: is required
          headers: {}
          x-apifox-ordering: 1
        '401':
          x-apifox-name: 未认证（Token 无效或缺失）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: UNAUTHENTICATED
                message: 需要提供有效的认证Token
          headers: {}
          x-apifox-ordering: 2
        '403':
          x-apifox-name: 禁止访问（例如：用户无权添加好友）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: FORBIDDEN
                message: 您的账户无权添加好友
          headers: {}
          x-apifox-ordering: 3
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 4
      security:
        - jwt: []
          x-apifox:
            required: true
            schemeGroups:
              - id: 2E3vUtLwdMRJwtG6NLo0V
                schemeIds:
                  - jwt
            use:
              id: 2E3vUtLwdMRJwtG6NLo0V
              configs:
                jwt:
                  authConfigs:
                    x-apifox:
                      token: >-
                        eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw
      x-apifox-folder: 好友
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-380410911-run
  /friends/{friendId}:
    get:
      summary: 获取单个好友详情
      deprecated: false
      description: ''
      tags:
        - 好友 (Friends)
      parameters:
        - name: friendId
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      responses:
        '200':
          x-apifox-name: 好友详情
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Friend'
          headers: {}
          x-apifox-ordering: 0
        '401':
          x-apifox-name: 未认证（Token 无效或缺失）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: UNAUTHENTICATED
                message: 需要提供有效的认证Token
          headers: {}
          x-apifox-ordering: 1
        '403':
          x-apifox-name: 禁止访问（例如：无权查看该私有好友）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: FORBIDDEN
                message: 您无权查看此好友
          headers: {}
          x-apifox-ordering: 2
        '404':
          x-apifox-name: 未找到好友
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: NOT_FOUND
                message: 未找到 ID 为 [friendId] 的好友
          headers: {}
          x-apifox-ordering: 3
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 4
      security:
        - jwt: []
          x-apifox:
            required: true
            schemeGroups:
              - id: 2E3vUtLwdMRJwtG6NLo0V
                schemeIds:
                  - jwt
            use:
              id: 2E3vUtLwdMRJwtG6NLo0V
              configs:
                jwt:
                  authConfigs:
                    x-apifox:
                      token: >-
                        eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw
      x-apifox-folder: 好友
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-380410912-run
    delete:
      summary: 删除好友
      deprecated: false
      description: ''
      tags:
        - 好友 (Friends)
      parameters:
        - name: friendId
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
      responses:
        '204':
          x-apifox-name: 删除成功，无返回内容
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
                x-apifox-orders: []
                x-apifox-ignore-properties: []
          headers: {}
          x-apifox-ordering: 0
        '401':
          x-apifox-name: 未认证（Token 无效或缺失）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: UNAUTHENTICATED
                message: 需要提供有效的认证Token
          headers: {}
          x-apifox-ordering: 1
        '403':
          x-apifox-name: 禁止访问（例如：无权删除他人的好友）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: FORBIDDEN
                message: 您没有权限删除此好友
          headers: {}
          x-apifox-ordering: 2
        '404':
          x-apifox-name: 未找到好友
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: NOT_FOUND
                message: 未找到 ID 为 [friendId] 的好友
          headers: {}
          x-apifox-ordering: 3
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 4
      security:
        - jwt: []
          x-apifox:
            required: true
            schemeGroups:
              - id: 2E3vUtLwdMRJwtG6NLo0V
                schemeIds:
                  - jwt
            use:
              id: 2E3vUtLwdMRJwtG6NLo0V
              configs:
                jwt:
                  authConfigs:
                    x-apifox:
                      token: >-
                        eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw
      x-apifox-folder: 好友
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-380410913-run
  /friends/requests:
    get:
      summary: 获取好友请求列表
      deprecated: false
      description: 获取当前用户收到的所有好友请求
      tags:
        - 好友 (Friends)
      parameters: []
      responses:
        '200':
          x-apifox-name: 成功的好友请求列表
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendRequestItem'
                uniqueItems: true
          headers: {}
          x-apifox-ordering: 0
        '401':
          x-apifox-name: 未认证（Token 无效或缺失）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: UNAUTHENTICATED
                message: 需要提供有效的认证Token
          headers: {}
          x-apifox-ordering: 1
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 2
      security:
        - jwt: []
          x-apifox:
            required: true
            schemeGroups:
              - id: 2E3vUtLwdMRJwtG6NLo0V
                schemeIds:
                  - jwt
            use:
              id: 2E3vUtLwdMRJwtG6NLo0V
              configs:
                jwt:
                  authConfigs:
                    x-apifox:
                      token: >-
                        eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw
      x-apifox-folder: 好友
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-380410914-run
  /friends/request:
    post:
      summary: 发送好友请求
      deprecated: false
      description: 向指定用户发送好友请求
      tags:
        - 好友 (Friends)
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendFriendRequestRequest'
            examples: {}
      responses:
        '200':
          x-apifox-name: 成功
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - 好友请求已发送
                x-apifox-orders:
                  - message
                x-apifox-ignore-properties: []
          headers: {}
          x-apifox-ordering: 0
        '400':
          x-apifox-name: 无效请求
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VALIDATION_ERROR
                message: 目标用户ID不能为空
          headers: {}
          x-apifox-ordering: 1
        '401':
          x-apifox-name: 未认证（Token 无效或缺失）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: UNAUTHENTICATED
                message: 需要提供有效的认证Token
          headers: {}
          x-apifox-ordering: 2
        '409':
          x-apifox-name: 资源冲突（例如：已经是好友或请求已存在）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: ALREADY_FRIENDS
                message: 你们已经是好友了
          headers: {}
          x-apifox-ordering: 3
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 4
      security:
        - jwt: []
          x-apifox:
            required: true
            schemeGroups:
              - id: 2E3vUtLwdMRJwtG6NLo0V
                schemeIds:
                  - jwt
            use:
              id: 2E3vUtLwdMRJwtG6NLo0V
              configs:
                jwt:
                  authConfigs:
                    x-apifox:
                      token: >-
                        eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw
      x-apifox-folder: 好友
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-380410915-run
  /friends/request/handle:
    post:
      summary: 处理好友请求
      deprecated: false
      description: 接受或拒绝好友请求
      tags:
        - 好友 (Friends)
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandleFriendRequestRequest'
      responses:
        '200':
          x-apifox-name: 好友请求处理成功
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - 好友请求已处理
                x-apifox-orders:
                  - message
                x-apifox-ignore-properties: []
          headers: {}
          x-apifox-ordering: 0
        '400':
          x-apifox-name: 无效请求
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VALIDATION_ERROR
                message: 请求ID不能为空
          headers: {}
          x-apifox-ordering: 1
        '401':
          x-apifox-name: 未认证（Token 无效或缺失）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: UNAUTHENTICATED
                message: 需要提供有效的认证Token
          headers: {}
          x-apifox-ordering: 2
        '404':
          x-apifox-name: 未找到好友请求
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: NOT_FOUND
                message: 未找到该好友请求
          headers: {}
          x-apifox-ordering: 3
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 4
      security:
        - jwt: []
          x-apifox:
            required: true
            schemeGroups:
              - id: 2E3vUtLwdMRJwtG6NLo0V
                schemeIds:
                  - jwt
            use:
              id: 2E3vUtLwdMRJwtG6NLo0V
              configs:
                jwt:
                  authConfigs:
                    x-apifox:
                      token: >-
                        eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw
      x-apifox-folder: 好友
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-380410916-run
  /friends/search:
    get:
      summary: 搜索用户
      deprecated: false
      description: 根据用户名、邮箱或手机号搜索用户
      tags:
        - 好友 (Friends)
      parameters:
        - name: query
          in: query
          description: 搜索关键词（用户名、邮箱或手机号）
          required: true
          schema:
            type: string
      responses:
        '200':
          x-apifox-name: 成功的搜索结果
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
                uniqueItems: true
          headers: {}
          x-apifox-ordering: 0
        '401':
          x-apifox-name: 未认证（Token 无效或缺失）
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: UNAUTHENTICATED
                message: 需要提供有效的认证Token
          headers: {}
          x-apifox-ordering: 1
        '500':
          x-apifox-name: 服务器内部错误
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: 服务器发生未知错误
          headers: {}
          x-apifox-ordering: 2
      security:
        - jwt: []
          x-apifox:
            required: true
            schemeGroups:
              - id: 2E3vUtLwdMRJwtG6NLo0V
                schemeIds:
                  - jwt
            use:
              id: 2E3vUtLwdMRJwtG6NLo0V
              configs:
                jwt:
                  authConfigs:
                    x-apifox:
                      token: >-
                        eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw
      x-apifox-folder: 好友
      x-apifox-status: released
      x-run-in-apifox: https://app.apifox.com/web/project/7359528/apis/api-380410917-run
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 业务错误码，便于前端处理
          examples:
            - VALIDATION_ERROR
        message:
          type: string
          description: 人类可读的错误信息
          examples:
            - 输入参数无效
        details:
          type: array
          description: （可选）详细的验证错误列表
          items:
            type: object
            properties:
              field:
                type: string
                examples:
                  - email
              issue:
                type: string
                examples:
                  - 邮箱格式不正确
            x-apifox-orders:
              - field
              - issue
            x-apifox-ignore-properties: []
      required:
        - code
        - message
      x-apifox-orders:
        - code
        - message
        - details
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
    Friend:
      type: object
      description: 好友信息
      properties:
        id:
          type: string
          description: 好友ID
        name:
          type: string
          description: 好友姓名
        email:
          type: string
          format: email
          description: 好友邮箱
        phone:
          type: string
          description: 好友手机号
        mutualFriends:
          type: integer
          description: 共同好友数量
        createdAt:
          type: string
          format: date-time
          description: 成为好友的时间
      required:
        - id
        - name
        - email
        - phone
      x-apifox-orders:
        - id
        - name
        - email
        - phone
        - mutualFriends
        - createdAt
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
    FriendRequest:
      type: object
      description: 好友请求（用于发送请求时）
      properties:
        targetUserId:
          type: string
          description: 目标用户ID
        message:
          type: string
          description: 附带消息（可选）
      required:
        - targetUserId
      x-apifox-orders:
        - targetUserId
        - message
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
    FriendRequestItem:
      type: object
      description: 好友请求项（接收到的请求）
      properties:
        id:
          type: string
          description: 请求ID
        senderId:
          type: string
          description: 发送者用户ID
        name:
          type: string
          description: 发送者姓名
        mutualFriends:
          type: integer
          description: 共同好友数量
        requestDate:
          type: string
          format: date-time
          description: 请求发送时间
        message:
          type: string
          description: 附带消息
      required:
        - id
        - senderId
        - name
        - mutualFriends
        - requestDate
      x-apifox-orders:
        - id
        - senderId
        - name
        - mutualFriends
        - requestDate
        - message
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
    SendFriendRequestRequest:
      type: object
      description: 发送好友请求的请求体
      properties:
        targetUserId:
          type: string
          description: 目标用户ID
        message:
          type: string
          description: 附带消息（可选）
      required:
        - targetUserId
      x-apifox-orders:
        - targetUserId
        - message
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
    HandleFriendRequestRequest:
      type: object
      description: 处理好友请求的请求体
      properties:
        requestId:
          type: string
          description: 好友请求ID
        accept:
          type: boolean
          description: true为接受，false为拒绝
      required:
        - requestId
        - accept
      x-apifox-orders:
        - requestId
        - accept
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
    SearchResult:
      type: object
      description: 用户搜索结果
      properties:
        id:
          type: string
          description: 用户ID
        name:
          type: string
          description: 用户姓名
        mutualFriends:
          type: integer
          description: 共同好友数量
        isFriend:
          type: boolean
          description: 是否已经是好友
        isPending:
          type: boolean
          description: 是否有待处理的好友请求
      required:
        - id
        - name
        - mutualFriends
        - isFriend
        - isPending
      x-apifox-orders:
        - id
        - name
        - mutualFriends
        - isFriend
        - isPending
      x-apifox-folder: ''
      x-apifox-ignore-properties: []
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-apifox:
        addTokenTo: header
        headerPrefix: Bearer
servers: []
security:
  - jwt: []
    x-apifox:
      required: true
      schemeGroups:
        - id: 2E3vUtLwdMRJwtG6NLo0V
          schemeIds:
            - jwt
      use:
        id: 2E3vUtLwdMRJwtG6NLo0V
        configs:
          jwt:
            authConfigs:
              x-apifox:
                token: >-
                  eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1IiwiaWF0IjoxNzY0NTkyMjE1LCJleHAiOjE3NjQ1OTU4MTV9.mwyJQo0q60DRrDS671GLziIZD2QEbVdpBPzmx4vdrxw